pyproject.toml
[tool.poetry]
name = "stopisha"
version = "0.0.0"
description = "Cross-Lingual Model for Hate Speech Detection on Twitter: A Case of Swahili and Swahili-English Slang"
authors = ["Andrew Kariuki <andrew.o.kariuki@gmail.com>"]
readme = "README.md"

[tool.black]
line-length = 88

[tool.isort]
profile= "hug"
src_paths=["stopisha","test"]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88

[tool.mypy]
python_version = 3.9
strict = true
follow_imports = "silent"
plugins = ["pydantic.mypy","sqlmypy"]

[[tool.mypy.overrides]]
module = "tests.*"
ignore_missing_imports = true
disallow_untyped_defs = true
allow_incomplete_defs = true
allow_untyped_calls = true

[tool.semantic_release]
version_variable = "pyproject.toml:version"
version_source="tag"
commit_subject="chore(release): release stopisha {version}"
branch = "main"
upload_to_release = true
patch_without_tag = false

[tool.bandit]
exclude_dirs = ["tests"]
tests = ["B201", "B301"]
skips = ["B101", "B601"]

[tool.poetry.dependencies]
python = "^3.11"
sphinx = "^6.1.3"
python-semantic-release = "^7.33.2"
fastapi-redis-cache = "^0.2.5"
flake8 = "^6.0.0"
greenlet = "^2.0.2"
fastapi = "^0.92.0"
autoflake = "^2.0.1"
bandit = "^1.7.4"
isort = "^5.12.0"
black = "^23.1.0"
pytest = "^7.2.2"
coverage = "^7.2.1"
preprocessor = "^1.1.3"
pandas = "^1.5.3"
xlsxwriter = "^3.0.8"
twython = "^3.9.1"
nltk = "^3.8.1"
tweet-preprocessor = "^0.6.0"
alphabet-detector = "^0.0.7"
tweepy = "^4.12.1"
argcomplete = "^2.1.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
